# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "AgentoGo API",
        "description": "API REST multi-tenant con autenticación, auditoría y gestión de negocios",
        "contact": {
            "name": "AgentoGo Support",
            "email": "support@agentogo.com"
        },
        "version": "0.3.0"
    },
    "servers": [
        {
            "url": "http://localhost:8000/api",
            "description": "Servidor de desarrollo"
        }
    ],
    "paths": {
...
        "/businesses": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Listar negocios del usuario",
                "operationId": "1706cbaf428057f56998d8e961c6862c",
                "responses": {
                    "200": {
                        "description": "Lista de negocios con paginación"
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "post": {
                "tags": [
                    "Business"
                ],
                "summary": "Crear un nuevo negocio",
                "operationId": "9a2f8ce933dcf23050126381b3f8c21a",
                "responses": {
                    "201": {
                        "description": "Negocio creado exitosamente"
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "El nombre del negocio es obligatorio."
                                        ]
                                    }
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
        "/businesses/{business_id}": {
            "get": {
                "tags": [
                    "Business"
                ],
                "summary": "Obtener detalles de un negocio",
                "operationId": "c4818a6e6952d5e73e47243dd81a95b6",
                "responses": {
                    "200": {
                        "description": "Detalles del negocio"
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "put": {
                "tags": [
                    "Business"
                ],
                "summary": "Actualizar un negocio",
                "operationId": "a972040fead5e0bac36fbac3e26dd5fe",
                "responses": {
                    "200": {
                        "description": "Negocio actualizado"
                    },
                    "422": {
                        "description": "Validación fallida",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/ValidationError"
                                },
                                "example": {
                                    "message": "The given data was invalid.",
                                    "errors": {
                                        "name": [
                                            "El nombre del negocio es obligatorio."
                                        ]
                                    }
                                }
                            }
                        }
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            },
            "delete": {
                "tags": [
                    "Business"
                ],
                "summary": "Eliminar un negocio",
                "operationId": "ece5bc88037ea0e5ccc6583aa0ce281e",
                "responses": {
                    "204": {
                        "description": "Negocio eliminado"
                    },
                    "403": {
                        "description": "Acceso denegado",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Error403"
                                },
                                "example": {
                                    "code": 403,
                                    "message": "No tiene permisos para realizar esta acción."
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "sanctum": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "schemas": {
            "Error403": {
                "properties": {
                    "code": {
                        "type": "integer",
                        "example": 403
                    },
                    "message": {
                        "type": "string",
                        "example": "No autorizado"
                    }
                },
                "type": "object",
                "example": {
                    "code": 403,
                    "message": "No tiene permisos para realizar esta acción."
                }
            },
            "ValidationError": {
                "properties": {
                    "message": {
                        "type": "string",
                        "example": "The given data was invalid."
                    },
                    "errors": {
                        "description": "Mapa de errores por campo",
                        "type": "object"
                    }
                },
                "type": "object",
                "example": {
                    "message": "The given data was invalid.",
                    "errors": {
                        "name": [
                            "El campo nombre es requerido."
                        ],
                        "email": [
                            "El email debe ser una dirección válida."
                        ]
                    }
                }
            }
        },
        "securitySchemes": {
            "sanctum": {
                "type": "http",
                "description": "Token de autenticación Sanctum",
                "bearerFormat": "token",
                "scheme": "bearer"
            }
        }
    },
...
}